utilの作成
