+++ lib/maiporarisu_util.dart
@@ -1,0 +1,34 @@
+class MaiporarisuUtil {
+  static String getTimeDifference(DateTime dateTime) {
+    var now = DateTime.now();
+    var difference = dateTime
+        .add(Duration(hours: dateTime.hour, minutes: dateTime.minute))
+        .difference(now);
+    var isNegative = difference.isNegative;
+
+    if (difference.inDays.abs() >= 365) {
+      var years = difference.inDays.abs() ~/ 365;
+      return isNegative ? '$years年前' : '$years年後';
+    } else if (difference.inDays.abs() >= 30) {
+      var months = difference.inDays.abs() ~/ 30;
+      return isNegative ? '$monthsヶ月前' : '$monthsヶ月後';
+    } else if (difference.inDays.abs() >= 7) {
+      var weeks = difference.inDays.abs() ~/ 7;
+      return isNegative ? '$weeks週間前' : '$weeks週間後';
+    } else if (difference.inDays.abs() > 0) {
+      return isNegative
+          ? '${difference.inDays.abs()}日前'
+          : '${difference.inDays}日後';
+    } else if (difference.inHours.abs() > 0) {
+      return isNegative
+          ? '${difference.inHours.abs()}時間前'
+          : '${difference.inHours}時間後';
+    } else if (difference.inMinutes.abs() > 0) {
+      return isNegative
+          ? '${difference.inMinutes.abs()}分前'
+          : '${difference.inMinutes}分後';
+    } else {
+      return '現在';
+    }
+  }
+}
